requirements.txt
langchain>=0.1.0
langchain-chroma>=0.0.1
langchain-groq>=0.0.3
langchain-huggingface>=0.0.1
langchain-core>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.18
python-dotenv>=1.0.0
gradio>=4.0.0
sentence-transformers>=2.2.2